+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+# [Unreleased]
+### Fixed
+- AS1/2 handle declaration of registers in certain cases
+
 ## [14.5.1] - 2021-11-20
 ### Fixed
 - AS 1/2 - do not use eval function on obfuscated increment/decrement